+++ libs/readPTU/readTTTRRecords-for-import.c
@@ -500,7 +500,7 @@ void calculate_g2_ring(FILE* filehandle, long long record_type, int end_of_heade
     uint64_t oldest_timetag;
     int channel = -1;
     
-    int i;  // loop indexing
+    size_t i;  // loop indexing
     int idx;  // index for histogram
     
     uint64_t delta;
@@ -546,6 +546,7 @@ void calculate_g2_ring(FILE* filehandle, long long record_type, int end_of_heade
     }
 }
 
+
 void calculate_g2(FILE* filehandle, long long record_type, int end_of_header, uint64_t *RecNum, uint64_t NumRecords, uint64_t RecNum_start, uint64_t RecNum_stop, uint64_t *time_vector, int *histogram, int nb_of_bins, int channel_start, int channel_stop)
 {
     /*
@@ -581,7 +582,7 @@ void calculate_g2(FILE* filehandle, long long record_type, int end_of_header, ui
     uint64_t oflcorrection = 0;
     uint64_t timetag = 0;
     int channel = -1;
-    int i = 0;
+    uint64_t i = 0;
     uint64_t correlation_window = 0;
     //    long next_print = 0;
     correlation_window = time_vector[nb_of_bins];
@@ -663,7 +664,7 @@ void calculate_g2(FILE* filehandle, long long record_type, int end_of_header, ui
             }
         }
         // remove stop photons which are out of the correlation window (pop)
-        for(i = 0; i < stop_corr_buff_length; i++) {
+        for(i = 0; i < (uint64_t) stop_corr_buff_length; i++) {
             if(stop_corr_buff_head->next->val < start_time) {
                 pop(stop_corr_buff_head, &stop_corr_buff_length);
             }
@@ -675,7 +676,7 @@ void calculate_g2(FILE* filehandle, long long record_type, int end_of_header, ui
         current = stop_corr_buff_head->next;
         while(current != NULL) {
             if (current->val - start_time < correlation_window) {
-                i = (int) (current->val - start_time) * nb_of_bins / correlation_window;
+                i = (uint64_t) (current->val - start_time) * nb_of_bins / correlation_window;
                 histogram[i] = histogram[i] + 1;
             }
             current = current->next;
@@ -711,7 +712,7 @@ void calculate_g2_fast(FILE* filehandle, long long record_type, int end_of_heade
     uint64_t oflcorrection = 0;
     uint64_t timetag = 0;
     int channel = -1;
-    int i = 0;
+    uint64_t i = 0;
     uint64_t correlation_window = 0;
     //    long next_print = 0;
     correlation_window = time_vector[nb_of_bins];
@@ -752,9 +753,8 @@ void calculate_g2_fast(FILE* filehandle, long long record_type, int end_of_heade
         // ADD DELAY TO HISTOGRAM
         // add occurence to result histogram if the delay is in the correlation window
         if (stop_time - start_time < correlation_window) {
-            i = (int) (stop_time - start_time) * nb_of_bins / correlation_window;
+            i = (uint64_t) (stop_time - start_time) * nb_of_bins / correlation_window;
             histogram[i] = histogram[i] + 1;
         }
     }
 }
-
